backend/requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bandit==1.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.7
cryptography==45.0.5
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.10.0
django-appconf==1.1.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-mptt==0.17.0
django-polymorphic==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-polymorphic==2.0.0
drf-spectacular==0.28.0
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==37.8.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
oauthlib==3.3.1
packaging==25.0
pilkit==3.0
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.52
pyaes==1.6.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
redis==6.4.0
referencing==0.36.2
reportlab==4.4.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
stevedore==5.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
